steward/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg[binary]==3.2.2
httpx==0.27.2
pyjwt[crypto]==2.9.0
fastapi-clerk-auth
